@@ -9,19 +9,17 @@
     ],
     "license": "MIT",
     "require": {
-        "php": ">=5.5.9",
-        "illuminate/support": "5.1.*",
-        "illuminate/database": "5.1.*"
+        "php": ">=7.4",
+        "illuminate/support": "*",
+        "illuminate/database": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": ">= 3.7.0",
-        "codeclimate/php-test-reporter": "dev-master"
+        "phpunit/phpunit": "8.*|9.*"
     },
     "autoload": {
-        "psr-0": {
-            "Goez\\TreeData": "src/",
-            "GoezTest\\TreeData": "tests/"
+        "psr-4": {
+            "Goez\\TreeData\\": "src/"
         }
     },
-    "minimum-stability": "dev"
+    "minimum-stability": "stable"
 }
